Add image credits to recipe pages and README.md
Restyled the recipes pages to contain image credits.
Added an arrow to the 'Home' link at the bottom of the recipes pages
and gave it a sleeker look for aesthetic purpose.
Restyled the links on the homepage to give them a sleeker look too.
Edited README.md markdown for a better look and added the image credits
there too.